Move backend/sftp.SplitShellArgs to backend/
