+++ internal/backend/shell_split.go
@@ -1,4 +1,4 @@
-package sftp
+package backend
 
 import (
 	"unicode"
