+++ internal/backend/shell_split_test.go
@@ -1,4 +1,4 @@
-package sftp
+package backend
 
 import (
 	"reflect"
